refactor(agent): split execution domain into graph, review sub-domains
