+++ packages/agent/src/execution/graph/file-lock.ts
