+++ packages/agent/src/execution/graph/index.ts
@@ -1,0 +1,15 @@
+export {
+  buildDependencyGraph,
+  completeTaskAndUnblockDependents,
+} from "./execution-graph";
+
+export {
+  resolveDispatchHybridRuntime,
+  assignAgentsToReadyTasks,
+  resolveFallbackAgentAssignment,
+  resolveWorkerRuntimeForTask,
+} from "./execution-assignment";
+
+export type { RunSupervisorToolDecision } from "./execution-assignment";
+
+export { FileLock } from "./file-lock";
