+++ packages/agent/src/execution/index.ts
@@ -1,41 +1,6 @@
 // Execution domain — DAG execution engine, graph building, agent assignment, review gate
 
-export type {
-  SupervisorDecision,
-  SummarizedHistory,
-  ExecutionPlan,
-  ExecutionPlanStep,
-  ExecutionSupervisorConfig,
-  ExecutionSupervisorResult,
-  StepOutcome,
-  DispatchContext,
-  DispatchExecutionInput,
-  DispatchOutput,
-  DispatchReviewDecision,
-  DispatchReviewInput,
-  DispatchTask,
-} from "./execution-types";
-
-export { ExecutionSupervisor } from "./execution-supervisor";
-
-export {
-  buildDependencyGraph,
-  completeTaskAndUnblockDependents,
-} from "./execution-graph";
-
-export {
-  resolveDispatchHybridRuntime,
-  assignAgentsToReadyTasks,
-  resolveFallbackAgentAssignment,
-  resolveWorkerRuntimeForTask,
-} from "./execution-assignment";
-
-export {
-  decideFailedStepAction,
-  reviewTaskResult,
-  sendReviewFeedback,
-  requestHandoffDocument,
-  rotateAgent,
-} from "./execution-review";
-
-export { FileLock } from "./file-lock";
+export * from "./execution-types";
+export * from "./execution-supervisor";
+export * from "./graph";
+export * from "./review";
