+++ packages/agent/test/integration/supervisor-worker-split.test.ts
@@ -3,7 +3,7 @@ import { Session } from "@openomni/session";
 import { TaskManager } from "../../src/task/manager";
 import { TaskStorage } from "../../src/task/storage";
 import { BuiltinAgentRegistry } from "../../src/agent/registry/registry";
-import { FileLock } from "../../src/execution/file-lock";
+import { FileLock } from "../../src/execution/graph/file-lock";
 import { RunWorker } from "../../src/worker/run/run-worker";
 import type { OrchestratorRunInput } from "../../src/worker/run/run-worker";
 import type { Sink, Tool } from "@openomni/protocol";
