+++ packages/agent/test/tools/dispatch.test.ts
@@ -5,7 +5,7 @@ import { BuiltinAgentRegistry } from "../../src/agent/registry/registry";
 import { AgentMessenger } from "../../src/agent/communication";
 import { TaskStorage } from "../../src/task/storage";
 import { Dispatch, type DispatchReviewInput } from "../../src/tools/dispatch";
-import { FileLock } from "../../src/execution/file-lock";
+import { FileLock } from "../../src/execution/graph/file-lock";
 import { TaskManager } from "../../src/task/manager";
 
 interface MockCall {
